+++ migrations/m171027_185418_user_id_index.php
@@ -6,14 +6,20 @@
  *
  */
 
-use yii\db\Migration;
+use humhub\components\Migration;
 
 class m171027_185418_user_id_index extends Migration
 {
     public function safeUp()
     {
         $this->execute('DELETE FROM calendar_entry_participant WHERE user_id NOT IN (SELECT id FROM user)');
-        $this->addForeignKey('fk-calendar-participant-user-id', 'calendar_entry_participant', 'user_id', 'user', 'id', 'CASCADE');
+
+        try {
+            // Seems to throw 'Foreign key constraint is incorrectly' formed error in some installations
+            $this->addFosreignKey('fk-calendar-participant-user-id', 'calendar_entry_participant', 'user_id', 'user', 'id', 'CASCADE');
+        } catch(\Throwable $e) {
+            Yii::error($e);
+        }
     }
 
     public function safeDown()