+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Chng: More stable event error handling 
 - Fix #167: Can not remove calendar type, once a calendar type is set
 - Fix #173: Event cannot be moved to another Space
+- Fix #175: Implemented fallback for failed `calendar_entry_participant` user foreign key
 
 1.0.6 (March 4, 2020)
 ---------------------